Deployment/requirements.txt
streamlit
pandas
numpy
matplotlib
plotly
scikit-learn
seaborn
Pillow